requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Flask-SQLAlchemy
SQLAlchemy<1.4
PyMySQL==1.0.2
gunicorn
flask_migrate
Pillow
cryptography
